modified:   src/librootfs/akaimer39lib/etc/init.d/power_off.sh 	modified:   src/librootfs/akaimer39lib/etc/init.d/power_on.sh 	modified:   src/librootfs/akaimer39lib/etc/init.d/setup.sh

+++ src/librootfs/akaimer39lib/etc/init.d/power_on.sh
@@ -10,9 +10,9 @@ sleep 3
 export HOME=/etc
 Signal=$(iwconfig wlan0 | grep Signal | cut -d "-" -f 2 | cut -d " " -f 1)
 	if [ $Signal -le 85 ] > /dev/null; then
-		if pidof camera.sh; then kill -2 `pidof camera.sh`; fi
-                if pgrep record_video; then kill -2 `pgrep record_video`;fi
-		if pgrep dropbear; then kill -2 `pgrep dropbear`; fi
+		if pidof camera.sh; then kill `pidof camera.sh`; fi
+                if pgrep record_video; then kill `pgrep record_video`;fi
+		if pgrep dropbear; then kill `pgrep dropbear`; fi
 		dropbear -R -B
 		ntpd -q -p time.windows.com
 		sleep 10
@@ -21,14 +21,14 @@ Signal=$(iwconfig wlan0 | grep Signal | cut -d "-" -f 2 | cut -d " " -f 1)
 		sleep 1
 		echo heartbeat > /sys/class/leds/g_led/trigger
 		echo `hwclock` > /etc/on_`hostname`.txt
-		umount -l /mnt
-		mount /dev/mmcblk0p1 /mnt
+		#umount -l /mnt
+		#mount /dev/mmcblk0p1 /mnt
 		find /mnt/ -name "*index" -exec rm {} \;
 		time=`date +%Y%m%d`
 		rsync -av --no-o --no-g --remove-source-files --password-file=/etc/.rsync /etc/on_`hostname`.txt root@$Server::video/ya.disk/Avtobus/$time/
 		rsync -av --no-o --no-g --remove-source-files --log-file=/etc/`hostname`.txt --password-file=/etc/.rsync /mnt/ root@$Server::video/oneday/
 		rsync -av --no-o --no-g --remove-source-files --password-file=/etc/.rsync /etc/`hostname`.txt root@$Server::video/ya.disk/Avtobus/$time/
-		find /mnt/[2-9]* -type d -delete
+		if [ -d /mnt/[2-9]* ]; then find /mnt/[2-9]* -type d -delete; fi
 		/etc/init.d/wifi_led.sh r_led on
 		/etc/init.d/wifi_led.sh g_led off
 	else
